requirements.txt
pandas
matplotlib
seaborn
wordcloud
bs4
sentence_transformers
numpy
faiss-cpu # or faiss-gpu if you have a compatible GPU
pickle
scikit-learn